Array de commands e builder de commands
