+++ file.js
@@ -24,9 +24,7 @@ const create = (item) => {
 
 const update = ({ _id, fields }) => {
     try{
-        const content = read(file)
-        let items = content.split('\r\n')
-        items = items.map(item => JSON.parse(item))
+        let items = get()
         Object.keys(fields).forEach((fKey) => { 
             items.map((item) => {
                 Object.keys(item).forEach((iKey) => {
@@ -35,7 +33,6 @@ const update = ({ _id, fields }) => {
                 })
             }) 
         })
-    
         let string = items.map(item => JSON.stringify(item)).join('\r\n')
         fs.writeFileSync(file, string)
         console.log(chalk.bold.green('Usuário atualizado com sucesso!'))
@@ -46,10 +43,8 @@ const update = ({ _id, fields }) => {
 
 const remove = (value) => {
     try{
-        const content = read(file)
-        let items = content.split('\r\n')
-        let filter = items.map(item => JSON.parse(item)).
-            filter(({ _id }) => _id !== value).
+        let items = get()
+        let filter = items.filter(({ _id }) => _id !== value).
             map(item => JSON.stringify(item)).
             join('\r\n')
 
@@ -58,6 +53,19 @@ const remove = (value) => {
     } catch(err){
         console.log(chalk.bold.red('Ouve um problema! Não foi possivel remover o usuario.'))
     }
+}
+
+const get = (id) => {
+    const content = read(file)
+    let items = content.split('\r\n')
+    items = items.map(item => JSON.parse(item))
+
+    if (id){
+        const item = items.find(({ _id }) => _id === id)
+        return item
+    }
+    
+    return items
 }
 
 const read = (file) => {
@@ -87,5 +95,6 @@ module.exports = {
     create, 
     update,
     remove,
-    getPath
+    getPath,
+    get
 }
